@@ -1,5 +1,5 @@
 module github.com/kuttiproject/workspace
 
-go 1.18
+go 1.22
 
-require github.com/kuttiproject/kuttilog v0.2.0
+require github.com/kuttiproject/kuttilog v0.2.1
